backend/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
pillow
python-dotenv
google-generativeai
sentence-transformers
faiss-cpu
requests
python-multipart

